move the classes to `user` package
